backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pandas==2.1.2
numpy==1.24.3
langchain==0.0.350
openai==1.3.0
faiss-cpu==1.7.4
python-dotenv==1.0.0 